fix(cli): enforce sane ranges on reassembly threshold flags
Closes the three process-gap deferrals the 3-pass adversarial review of
the snaplen / small-segment sub-cycle logged in STATE.md.

- Drift 6 — the reassembly threshold flags accepted any value, so an
  operator could silently configure a nonsensical detector (e.g.
  `--small-segment-max-bytes 100000`, which makes every TCP segment
  "small"). The flags now carry clap `value_parser` ranges, enforced at
  parse time, matching the Snort ranges the config field docs cite:
  `--overlap-threshold` 0–255, `--small-segment-threshold` 0–2048,
  `--small-segment-max-bytes` 0–2048. `small_segment_max_bytes` is now
  `u16` (a payload-size cutoff never needs more). `--out-of-window-threshold`
  stays unbounded — no NIDS prior art gives it a defensible ceiling.

- Drift 7 — the threshold-boundary test pattern existed only for the
  small-segment detector. Added exactly-threshold negative tests for
  the overlap and out-of-window detectors, so a regression from `>` to
  `>=` is caught for all three.

- Drift 8 — documented (in `Cargo.toml` and `src/decoder.rs`) that the
  decoder's strict→lax `SliceError::Len` discrimination is coupled to
  the etherparse 0.16 API; `"0.16"` already excludes 0.17 via Cargo's
  caret semantics, and the IPv6/corrupt-packet tests act as contract
  tests for a future bump.

Tests: `--*-threshold` / `--small-segment-max-bytes` out-of-range
values are rejected at parse time; overlap and out-of-window
exactly-threshold runs stay silent.

+++ src/cli.rs
@@ -75,21 +75,24 @@ pub struct Cli {
     pub reassembly_memcap: usize,
 
     /// Override the overlapping-segment anomaly threshold (default: 50).
-    /// Per flow direction; see LESSON-P2.05 in the reassembly config.
-    #[arg(long, global = true)]
+    /// Per flow direction; accepted range 0–255 (Snort's
+    /// `overlap_limit` range). See LESSON-P2.05 in the reassembly config.
+    #[arg(long, global = true, value_parser = clap::value_parser!(u32).range(0..=255))]
     pub overlap_threshold: Option<u32>,
 
     /// Override the small-segment anomaly threshold (default: 100).
     /// Length of a consecutive run of undersized segments, per flow
-    /// direction, above which the anomaly fires. See LESSON-P2.05.
-    #[arg(long, global = true)]
+    /// direction, above which the anomaly fires. Accepted range 0–2048
+    /// (Snort's `small_segments` count range). See LESSON-P2.05.
+    #[arg(long, global = true, value_parser = clap::value_parser!(u32).range(0..=2048))]
     pub small_segment_threshold: Option<u32>,
 
     /// Override the small-segment payload-size cutoff in bytes
     /// (default: 16). A segment shorter than this counts as "small";
-    /// 0 disables small-segment detection. See LESSON-P2.05.
-    #[arg(long, global = true)]
-    pub small_segment_max_bytes: Option<usize>,
+    /// 0 disables small-segment detection. Accepted range 0–2048
+    /// (Snort's `small_segments` size range). See LESSON-P2.05.
+    #[arg(long, global = true, value_parser = clap::value_parser!(u16).range(0..=2048))]
+    pub small_segment_max_bytes: Option<u16>,
 
     /// Override the ports exempt from small-segment detection
     /// (default: 23,513 — telnet, rlogin). Comma-separated; a flow is
